requirements.txt
asgiref==3.8.1
cffi==1.16.0
configparser==7.0.0
cryptography==42.0.7
Django==4.2.13
djangorestframework==3.15.1
mysql-connector-python==8.4.0
mysqlclient==2.2.4
pycparser==2.22
PyMySQL==1.1.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
